+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/netty/responsewriter/NettyResponseWriter.java
@@ -28,6 +28,7 @@ import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandlerContext;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 /*
@@ -36,7 +37,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class NettyResponseWriter {
 
-  private final ChannelHandlerContext ctx;
+  @Getter private final ChannelHandlerContext ctx;
   private final MockServerConfiguration configuration;
 
   public NettyResponseWriter(MockServerConfiguration configuration, ChannelHandlerContext ctx) {